+++ Tests/AmplitudeTests/Utilities/IdentifyInterceptorTests.swift
@@ -11,6 +11,7 @@ final class IdentifyInterceptorTests: XCTestCase {
     private var interceptor: TestIdentifyInterceptor!
     private var configuration: Configuration!
     private var pipeline: EventPipeline!
+    private var mockPathCreation: MockPathCreation!
 
     override func setUp() {
         super.setUp()
@@ -21,9 +22,12 @@ final class IdentifyInterceptorTests: XCTestCase {
             apiKey: "testApiKey",
             storageProvider: storage,
             identifyStorageProvider: identifyStorage,
-            identifyBatchIntervalMillis: identifyBatchIntervalMillis
+            identifyBatchIntervalMillis: identifyBatchIntervalMillis,
+            offline: NetworkConnectivityCheckerPlugin.Disabled
         )
         let amplitude = Amplitude(configuration: configuration)
+        mockPathCreation = MockPathCreation()
+        amplitude.add(plugin: NetworkConnectivityCheckerPlugin(pathCreation: mockPathCreation))
         httpClient = FakeHttpClient(configuration: configuration)
         pipeline = EventPipeline(amplitude: amplitude)
         pipeline.httpClient = httpClient